Add vessel_identifier in pending alerts
