+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/alerts/PendingAlert.kt
@@ -1,5 +1,6 @@
 package fr.gouv.cnsp.monitorfish.domain.entities.alerts
 
+import fr.gouv.cnsp.monitorfish.domain.entities.VesselIdentifier
 import fr.gouv.cnsp.monitorfish.domain.entities.alerts.type.AlertType
 import java.time.ZonedDateTime
 import java.util.*
@@ -9,6 +10,7 @@ class PendingAlert(val id: Int? = null,
                    val internalReferenceNumber: String? = null,
                    val externalReferenceNumber: String? = null,
                    val ircs: String? = null,
+                   val vesselIdentifier: VesselIdentifier,
                    val creationDate: ZonedDateTime,
                    val tripNumber: Int? = null,
                    val value: AlertType)
