+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/database/entities/PendingAlertEntity.kt
@@ -1,16 +1,26 @@
 package fr.gouv.cnsp.monitorfish.infrastructure.database.entities
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.vladmihalcea.hibernate.type.array.EnumArrayType
+import com.vladmihalcea.hibernate.type.basic.PostgreSQLEnumType
 import com.vladmihalcea.hibernate.type.json.JsonBinaryType
+import fr.gouv.cnsp.monitorfish.domain.entities.VesselIdentifier
 import fr.gouv.cnsp.monitorfish.domain.entities.alerts.PendingAlert
 import fr.gouv.cnsp.monitorfish.domain.entities.alerts.type.AlertType
 import org.hibernate.annotations.Type
 import org.hibernate.annotations.TypeDef
+import org.hibernate.annotations.TypeDefs
 import java.time.ZonedDateTime
 import javax.persistence.*
 
 @Entity
-@TypeDef(name = "jsonb", typeClass = JsonBinaryType::class)
+@TypeDefs(
+        TypeDef(name = "jsonb", typeClass = JsonBinaryType::class),
+        TypeDef(
+                name = "pgsql_enum",
+                typeClass = PostgreSQLEnumType::class
+        )
+)
 @Table(name = "pending_alerts")
 data class PendingAlertEntity(
         @Id
@@ -26,6 +36,10 @@ data class PendingAlertEntity(
         val externalReferenceNumber: String? = null,
         @Column(name = "ircs", nullable = false)
         val ircs: String? = null,
+        @Column(name = "vessel_identifier")
+        @Enumerated(EnumType.STRING)
+        @Type(type = "pgsql_enum")
+        val vesselIdentifier: VesselIdentifier,
         @Column(name = "creation_date", nullable = false)
         val creationDate: ZonedDateTime,
         @Column(name = "trip_number")
@@ -41,6 +55,7 @@ data class PendingAlertEntity(
                     internalReferenceNumber = internalReferenceNumber,
                     externalReferenceNumber = externalReferenceNumber,
                     ircs = ircs,
+                    vesselIdentifier = vesselIdentifier,
                     creationDate = creationDate,
                     tripNumber = tripNumber,
                     value = mapper.readValue(value, AlertType::class.java)
@@ -53,6 +68,7 @@ data class PendingAlertEntity(
                         internalReferenceNumber = alert.internalReferenceNumber,
                         externalReferenceNumber = alert.externalReferenceNumber,
                         ircs = alert.ircs,
+                        vesselIdentifier = alert.vesselIdentifier,
                         creationDate = alert.creationDate,
                         tripNumber = alert.tripNumber,
                         value = mapper.writeValueAsString(alert.value))
