+++ backend/src/main/resources/db/migration/internal/V0.114__Update_pending_alerts_table.sql
@@ -1,0 +1,2 @@
+DELETE FROM public.pending_alerts;
+ALTER TABLE public.pending_alerts ADD COLUMN vessel_identifier vessel_identifier NOT NULL;
